fix(js): refactor document parts to export via jsonschema
