+++ genkit-tools/common/src/types/document.ts
@@ -22,14 +22,29 @@ import z from 'zod';
 const EmptyPartSchema = z.object({
   text: z.never().optional(),
   media: z.never().optional(),
+  toolRequest: z.never().optional(),
+  toolResponse: z.never().optional(),
+  data: z.unknown().optional(),
+  metadata: z.record(z.unknown()).optional(),
+  custom: z.record(z.unknown()).optional(),
 });
 
+/**
+ * Zod schema for a text part.
+ */
 export const TextPartSchema = EmptyPartSchema.extend({
-  /** The text of the document. */
+  /** The text of the message. */
   text: z.string(),
 });
+
+/**
+ * Text part.
+ */
 export type TextPart = z.infer<typeof TextPartSchema>;
 
+/**
+ * Zod schema of a media part.
+ */
 export const MediaPartSchema = EmptyPartSchema.extend({
   media: z.object({
     /** The media content type. Inferred from data uri if not provided. */
@@ -38,13 +53,86 @@ export const MediaPartSchema = EmptyPartSchema.extend({
     url: z.string(),
   }),
 });
+
+/**
+ * Media part.
+ */
 export type MediaPart = z.infer<typeof MediaPartSchema>;
 
-export const PartSchema = z.union([TextPartSchema, MediaPartSchema]);
-export type Part = z.infer<typeof PartSchema>;
+/**
+ * Zod schema of a tool request part.
+ */
+export const ToolRequestPartSchema = EmptyPartSchema.extend({
+  /** A request for a tool to be executed, usually provided by a model. */
+  toolRequest: z.object({
+    /** The call id or reference for a specific request. */
+    ref: z.string().optional(),
+    /** The name of the tool to call. */
+    name: z.string(),
+    /** The input parameters for the tool, usually a JSON object. */
+    input: z.unknown().optional(),
+  }),
+});
+
+/**
+ * Tool part.
+ */
+export type ToolRequestPart = z.infer<typeof ToolRequestPartSchema>;
+
+/**
+ * Zod schema of a tool response part.
+ */
+export const ToolResponsePartSchema = EmptyPartSchema.extend({
+  /** A provided response to a tool call. */
+  toolResponse: z.object({
+    /** The call id or reference for a specific request. */
+    ref: z.string().optional(),
+    /** The name of the tool. */
+    name: z.string(),
+    /** The output data returned from the tool, usually a JSON object. */
+    output: z.unknown().optional(),
+  }),
+});
+
+/**
+ * Tool response part.
+ */
+export type ToolResponsePart = z.infer<typeof ToolResponsePartSchema>;
+
+/**
+ * Zod schema of a data part.
+ */
+export const DataPartSchema = EmptyPartSchema.extend({
+  data: z.unknown(),
+});
+
+/**
+ * Data part.
+ */
+export type DataPart = z.infer<typeof DataPartSchema>;
+
+/**
+ * Zod schema of a custom part.
+ */
+export const CustomPartSchema = EmptyPartSchema.extend({
+  custom: z.record(z.any()),
+});
+
+/**
+ * Custom part.
+ */
+export type CustomPart = z.infer<typeof CustomPartSchema>;
+
+// Disclaimer: genkit/js/ai/document.ts defines the following schema, type pair
+// as PartSchema and Part, respectively. genkit-tools cannot retain those names
+// due to it clashing with similar schema in model.ts, and genkit-tools
+// exporting all types at root. We use a different name here and updated
+// coresponding the imports.
+export const DocumentPartSchema = z.union([TextPartSchema, MediaPartSchema]);
+export type DocumentPart = z.infer<typeof DocumentPartSchema>;
 
 export const DocumentDataSchema = z.object({
-  content: z.array(PartSchema),
+  content: z.array(DocumentPartSchema),
   metadata: z.record(z.string(), z.any()).optional(),
 });
 export type DocumentData = z.infer<typeof DocumentDataSchema>;
