+++ genkit-tools/common/src/types/index.ts
@@ -19,6 +19,7 @@ export { GenkitErrorData } from '../types/error';
 export * from './action';
 export * from './analytics';
 export * from './apis';
+export * from './document';
 export * from './env';
 export * from './eval';
 export * from './evaluators';
